Added new feature to App, added tests for model classes on client side

+++ java/tenmo-client/src/main/java/com/techelevator/tenmo/models/Accounts.java
@@ -12,7 +12,7 @@ public class Accounts {
 		
 	}
 	
-	public Accounts(int accountId, int userId, BigDecimal balance) {
+	public Accounts(long accountId, long userId, BigDecimal balance) {
 		super();
 		this.accountId = accountId;
 		this.userId = userId;
